some small refactors
